Power-Up
-include a power-up that allows for mid-air boosts.
-new folder named Booster
-create a new material named Booster Mat.
-Mark the collider as a trigger,
-new C# script named Booster in the corresponding folder
-four public variables used to configure it: offset,
rotationVelocity,recycleOffset, spawnChance
-One way to make the spawning work is by requesting a booster placement
each time a platform is recycled. Then it's up to the booster itself
whether it'll be placed. We'll add a method named SpawnIfAvailable to
Booster for this.
-We then add a variable to PlatformManager to which we assign Booster.
Inside the Recycle method, we'll call its PlaceIfAvailable method after
we've determined the new platform's position.
-pdate the SpawnIfAvailable method so it activates and positions the
booster.
- Booster must begin deactivated and must also deactivate when the game
ends.
-to make the booster spin and recycle, we have to add an Update method
to it.
-boost velocity